panopticon/requirements.txt
asgiref==3.8.1
av==13.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==5.1.5
django-environ==0.12.0
dnspython==2.7.0
google-crc32c==1.6.0
idna==3.10
ifaddr==0.2.0
pillow==11.1.0
psycopg==3.2.11
pycparser==2.22
pyee==12.1.1
PyJWT==2.10.1
pylibsrtp==0.11.0
pyOpenSSL==25.0.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
